+++ README.md
@@ -61,6 +61,15 @@ CC1101SPIDevice = 0
 CC1101OutputPower = 0xC6
 ```
 
+For the original raw 433 MHz ASK/OOK transmitter module, use:
+
+```ini
+TXGPIO = 4
+RFBackend = raw_433
+```
+
+The older `RFBackend = gpio` value is still accepted as an alias for `raw_433`.
+
 
 ![Full Picture](documentation/Full%20Assembly.jpg)<br/>
 ![Pi Connection](documentation/Connection.jpg)<br/>
