+++ defaultConfig.conf
@@ -26,9 +26,10 @@ SendRepeat = 2
 TXGPIO = 4
 
 # (Optional) RF backend used for transmission. Supported values:
-#   gpio   - original 3-pin 433.42 MHz transmitter
-#   cc1101 - CC1101 modules such as the E07-M1101D-SMA, with GDO0 wired to TXGPIO
-RFBackend = gpio
+#   raw_433 - raw 433.42 MHz ASK/OOK transmitter driven from TXGPIO
+#   cc1101  - CC1101 modules such as the E07-M1101D-SMA, with GDO0 wired to TXGPIO
+# Legacy value "gpio" is accepted as an alias for raw_433.
+RFBackend = raw_433
 
 # (Optional) CC1101 backend settings used only when RFBackend = cc1101.
 # The E07-M1101D-SMA is a 3.3V SPI CC1101 module. Do not power it from 5V.
