+++ docs/superpowers/plans/2026-06-05-e07-m1101d-sma-cc1101.md
@@ -27,10 +27,10 @@
 - Create: `tests/test_rf_backend.py`
 - Modify: `operateShutters.py`
 
-- [ ] Write a test that instantiates `Shutter` with `RFBackend = gpio` and patches `_sendWave_gpio` to assert it is called.
-- [ ] Write a test that instantiates `Shutter` with `RFBackend = cc1101`, injects a fake CC1101 module, patches `_sendWave_gpio`, and asserts the radio is configured and `asynchronous_transmission()` wraps the waveform call.
+- [ ] Write a test that instantiates `Shutter` with `RFBackend = raw_433` and patches the raw 433 transmitter to assert it is called.
+- [ ] Write a test that instantiates `Shutter` with `RFBackend = cc1101`, injects a fake CC1101 module, patches the raw 433 transmitter, and asserts the radio is configured and `asynchronous_transmission()` wraps the waveform call.
 - [ ] Run `python3 -m unittest tests.test_rf_backend -v` and verify it fails because backend dispatch does not exist yet.
-- [ ] Extract the existing pigpio transmit block into `_sendWave_pigpio`, keep the existing lgpio path, add `_sendWave_gpio`, and add `_sendWave_cc1101`.
+- [ ] Extract the existing pigpio and lgpio transmit paths into a raw 433 transmitter backend, then add a CC1101 backend that wraps that waveform path.
 - [ ] Re-run backend tests and verify they pass.
 
 ### Task 3: Dependencies And Docs
