+++ tests/test_rf_backend_factory.py
@@ -2,8 +2,9 @@ import types
 import unittest
 
 from cc1101_backend import CC1101Transmitter
-from gpio_backend import GPIOTransmitter
+from raw_433_backend import Raw433Transmitter
 from rf_backend import create_transmitter
+from rf_backend import get_backend_name
 
 
 class FakeRadio:
@@ -16,7 +17,18 @@ class FakePigpio:
 
 
 class BackendFactoryTest(unittest.TestCase):
-    def test_creates_gpio_transmitter_by_default(self):
+    def test_creates_raw_433_transmitter_by_default(self):
+        config = types.SimpleNamespace(RFBackend="raw_433", TXGPIO=4)
+
+        transmitter = create_transmitter(
+            config,
+            is_pi5=False,
+            pigpio_module=FakePigpio(),
+        )
+
+        self.assertIsInstance(transmitter, Raw433Transmitter)
+
+    def test_accepts_legacy_gpio_backend_alias(self):
         config = types.SimpleNamespace(RFBackend="gpio", TXGPIO=4)
 
         transmitter = create_transmitter(
@@ -25,9 +37,10 @@ class BackendFactoryTest(unittest.TestCase):
             pigpio_module=FakePigpio(),
         )
 
-        self.assertIsInstance(transmitter, GPIOTransmitter)
+        self.assertEqual("raw_433", get_backend_name(config))
+        self.assertIsInstance(transmitter, Raw433Transmitter)
 
-    def test_creates_cc1101_transmitter_wrapping_gpio_transmitter(self):
+    def test_creates_cc1101_transmitter_wrapping_raw_433_transmitter(self):
         config = types.SimpleNamespace(
             RFBackend="cc1101",
             TXGPIO=4,
@@ -45,7 +58,7 @@ class BackendFactoryTest(unittest.TestCase):
         )
 
         self.assertIsInstance(transmitter, CC1101Transmitter)
-        self.assertIsInstance(transmitter.waveform_transmitter, GPIOTransmitter)
+        self.assertIsInstance(transmitter.waveform_transmitter, Raw433Transmitter)
 
     def test_rejects_unknown_backend(self):
         config = types.SimpleNamespace(RFBackend="other", TXGPIO=4)
